one file per class and fix get_html to get_html()
